+++ src/client/components/admin/user-modal/index.ts
@@ -1,6 +1,5 @@
 import Vue from "vue";
-import { post, put, apiurl, del } from "../../../helpers";
-import { state } from "../../../state";
+import { state, post, put, apiurl, del } from "../../../helpers";
 
 export const UserModal = Vue.extend({
     template: require("./user-modal.html"),
