+++ src/client/components/rider/rider-row/index.ts
@@ -1,5 +1,5 @@
 import Vue, { PropType } from "vue";
-import { state } from "../../../state";
+import { state } from "../../../helpers";
 import { IRider } from "../../../../shared";
 
 export const RiderRow = Vue.extend({
