+++ src/client/helpers/session.ts
@@ -1,5 +1,5 @@
-import { State, setAjaxToken } from "./helpers";
-import { emptyUser } from "../shared";
+import { State, setAjaxToken } from ".";
+import { emptyUser } from "../../shared";
 import VueRouter from "vue-router";
 const localStorage = require("localStorage");
 
@@ -17,6 +17,12 @@ export function signOut(router: VueRouter) {
     localStorage.setItem("session", "");
     state.set({ loggedIn: false, token: "" });
     router.push("/login");
+}
+
+export function updateJWT(newToken: string) {
+    localStorage.setItem("session", JSON.stringify({ newToken }));
+    state.set({ token: newToken });
+    setAjaxToken(newToken);
 }
 
 export function checkForSession() {
