+++ src/client/router.ts
@@ -1,7 +1,6 @@
 import VueRouter from "vue-router";
 import { HomePage, LoginPage, ShowClassPage, RiderPage, AdminPage } from "./pages";
-import { state, signOut } from "./state";
-import { post, apiurl } from "./helpers";
+import { state, signOut, post, apiurl } from "./helpers";
 
 export const router = new VueRouter({
     linkActiveClass: "is-active",
