+++ src/client/test/state.test.ts
@@ -1,6 +1,5 @@
-import { state, checkForSession } from "../state";
 import { assert } from "chai";
-import { emptyUser, User, Show, ShowClass, Rider } from "../../shared";
+import { state, checkForSession, emptyUser, User, Show, ShowClass, Rider } from "../../shared";
 const localStorage = require("localStorage");
 
 const testToken = "12345";
